+++ impl/maven-core/src/main/java/org/apache/maven/project/artifact/DefaultProjectArtifactsCache.java
@@ -27,13 +27,12 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
 import org.apache.maven.RepositoryUtils;
 import org.apache.maven.artifact.Artifact;
-import org.apache.maven.impl.cache.RefConcurrentMap;
+import org.apache.maven.impl.cache.Cache;
 import org.apache.maven.lifecycle.LifecycleExecutionException;
 import org.apache.maven.lifecycle.internal.SetWithResolutionResult;
 import org.apache.maven.project.MavenProject;
@@ -158,8 +157,8 @@ public class DefaultProjectArtifactsCache implements ProjectArtifactsCache {
         }
     }
 
-    protected final Map<Key, CacheRecord> cache = RefConcurrentMap.softMap();
-    protected final Map<Key, Key> keys = RefConcurrentMap.softMap();
+    protected final Cache<Key, CacheRecord> cache = Cache.newCache(Cache.ReferenceType.SOFT);
+    protected final Cache<Key, Key> keys = Cache.newCache(Cache.ReferenceType.SOFT);
 
     @Override
     public Key createKey(
@@ -201,18 +200,14 @@ public class DefaultProjectArtifactsCache implements ProjectArtifactsCache {
             throw new IllegalArgumentException("projectArtifacts must implement ArtifactsSetWithResult");
         }
 
-        CacheRecord record = new CacheRecord(artifacts);
-        cache.put(key, record);
-        return record;
+        return cache.computeIfAbsent(key, k -> new CacheRecord(artifacts));
     }
 
     @Override
     public CacheRecord put(Key key, LifecycleExecutionException exception) {
         Objects.requireNonNull(exception, "exception cannot be null");
         assertUniqueKey(key);
-        CacheRecord record = new CacheRecord(exception);
-        cache.put(key, record);
-        return record;
+        return cache.computeIfAbsent(key, k -> new CacheRecord(exception));
     }
 
     protected void assertUniqueKey(Key key) {
